--- client/src/components/modules/Constants.js
@@ -1,86 +1,0 @@
-const timeZones = [
-  {
-    value: "GMT-4",
-    timezone: "Eastern Daylight Time - Boston (GMT-4)"
-  },
-  {
-    value: "GMT-5",
-    timezone: "Central Daylight Time - Chicago (GMT-5)"
-  },
-  {
-    value: "GMT-6",
-    timezone: "Mountain Daylight Time - Denver (GMT-6)"
-  },
-  {
-    value: "GMT-7",
-    timezone: "Mountain Standard Time - Phoenix (GMT-7)"
-  },
-  {
-    value: "GMT-7",
-    timezone: "Pacific Daylight Time - Los Angeles (GMT-7)"
-  },
-  {
-    value: "GMT-8",
-    timezone: "Alaska Daylight Time - Anchorage (GMT-8)"
-  },
-  {
-    value: "GMT-10",
-    timezone: "Hawaii-Aleutian Standard Time - Honolulu (GMT-10)"
-  },
-]
-
-export const subjects = [
-  {
-    value: "English",
-    label: "English"
-  },
-  {
-    value: "History",
-    label: "History"
-  },
-  {
-    value: "Science",
-    label: "Science"
-  },
-  {
-    value: "Math",
-    label: "Math"
-  },
-  {
-    value: "Physics",
-    label: "Physics"
-  },
-  {
-    value: "Chemistry",
-    label: "Chemistry"
-  },
-  {
-    value: "Biology",
-    label: "Biology"
-  },
-  {
-    value: "College Prep - SATs, ACTs",
-    label: "College Prep - SATs, ACTs",
-  },
-  {
-    value: "College Prep - Essays",
-    label: "College Prep - Essays"
-  },
-  {
-    value: "Writing",
-    label: "Writing",
-  },
-  {
-    value: "Economics",
-    label: "Economics"
-  },
-  {
-    value: "Computer Science",
-    label: "Computer Science"
-  }
-]
-
-export const tags = ["Women in STEM", "FGLI", "SPED: Special Education", "ASL: American Sign Language", "LGBTQ + ", "English as a Second Language", "Underrepresented Minority in STEM(African Americans, American Indians / Alaska Natives, and Latinos)", "Early Childhood Learning", "High School", "Middle School", "Elementary School", "Preschool"];
-
-
-export default timeZones;
