+++ client/src/components/pages/SignIn.js
@@ -1,6 +1,7 @@
 import React, { Component } from "react";
 import 'bootstrap/dist/css/bootstrap.min.css';
 import "../../utilities.css";
+import { theme } from "../Constants.js";
 
 import { Col, Row, Button, Container } from 'react-bootstrap'
 import Form from "react-bootstrap/Form";
@@ -23,38 +24,6 @@ const withNavigate = (Component) => {
     return <Component {...props} navigate={navigate} />;
   }
 }
-
-const theme = {
-  colors: {
-    blue: '#00568C',
-    yellow: '#F2BE32',
-    white: '#ffffff',
-    darkblue: '#003c61',
-  },
-  fonts: {
-    sans: 'Muli, sans-serif',
-  },
-  fontWeights: {
-    light: 300,
-    normal: 600,
-    bold: 700,
-  },
-  fontSizes: [
-    12, 16, 24, 36, 48, 72
-  ],
-  space: [
-    0,
-    4,
-    8,
-    16,
-    32,
-    64,
-    128,
-    140,
-    256,
-  ]
-}
-
 
 class SignIn extends Component {
   // set context
